More functionalities

+++ E_Office/src/main/java/com/kavit/dao/DepartmentRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.kavit.model.Department;
 
-public interface DepartmentRepo extends JpaRepository<Department, Integer>{
+public interface DepartmentRepo extends JpaRepository<Department, Long>{
 
 }
